requirements.txt
plaid-python
pyyaml
twilio
numpy 